refactor: llamalend markets helpers
