+++ tests/cypress/e2e/llamalend/llamalend-markets.cy.ts
@@ -1,17 +1,28 @@
-import { maxBy, mean, minBy, sum } from 'lodash'
+import { sum } from 'lodash'
 import { LlamaMarketColumnId } from '@/llamalend/features/market-list/columns/columns.enum'
 import type { GetMarketsResponse } from '@curvefi/prices-api/llamalend'
-import { oneOf, shuffle, type TokenType } from '@cy/support/generators'
+import { oneOf, shuffle } from '@cy/support/generators'
 import {
   closeDrawer,
   expandFirstRowOnMobile,
+  getTableCellAssets,
   openDrawer,
   openFilters,
-  withFiltersPopover,
   withFilterChips,
-  getTableCellAssets,
+  withFiltersPopover,
 } from '@cy/support/helpers/data-table.helpers'
 import { Chain, HighTVLAddress, HighUtilizationAddress } from '@cy/support/helpers/lending-mocks'
+import {
+  checkChainSelection,
+  checkCoinSelection,
+  checkLineGraphColor,
+  checkTableFilterButtonGroupSelection,
+  enableGraphColumn,
+  filterByMarketType,
+  getOneColumnMiddleRangeValue,
+  typeFilterInput,
+  visitAndWait,
+} from '@cy/support/helpers/llamalend/llamalend-markets'
 import { setupLlamalendListMocks } from '@cy/support/helpers/llamalend/market-list-mocks'
 import {
   assertInViewport,
@@ -21,8 +32,7 @@ import {
   oneViewport,
   RETRY_IN_CI,
 } from '@cy/support/ui'
-import { fromEntries, objectKeys, range, recordEntries, recordValues, repeat } from '@primitives/objects.utils'
-import { serializeListFilter } from '@ui-kit/shared/ui/DataTable/filters'
+import { objectKeys, range, recordValues, repeat } from '@primitives/objects.utils'
 import { LlamaMarketType, LlamaMarketVersion, MarketRateType } from '@ui-kit/types/market'
 
 const wstEthMarket = '0x100dAa78fC509Db39Ef7D04DE0c1ABD299f4C6CE' as const
@@ -222,8 +232,8 @@ testCases.forEach(([width, height, breakpoint]) => {
 
     it(`should allow filtering by token`, () => {
       openFilters()
-      checkCoinSelection('collateral')
-      checkCoinSelection('borrowed')
+      checkCoinSelection(vaultData, 'collateral')
+      checkCoinSelection(vaultData, 'borrowed')
     })
 
     it('should allow filtering by using a range inputs', () => {
@@ -282,98 +292,5 @@ testCases.forEach(([width, height, breakpoint]) => {
       cy.url().should('not.include', 'isFavorite=')
       cy.get(`[data-testid^="data-table-row"]`).should('have.length.above', 1)
     })
-
-    function checkLineGraphColor(type: MarketRateType, color: string) {
-      // the graphs are lazy loaded, so we need to scroll to them first before checking the color
-      cy.get(`[data-testid="line-graph-${type}"]:visible`).first().scrollIntoView()
-      cy.get(`[data-testid="line-graph-${type}"] path`, LOAD_TIMEOUT).first().should('have.attr', 'stroke', color)
-    }
-
-    function checkCoinSelection(type: TokenType) {
-      const symbol = oneOf(...vaultData.ethereum.data.map(d => d[`${type}_token`].symbol))
-      const columnId = `assets_${type}_symbol`
-      cy.get(`[data-testid="multi-select-filter-${columnId}"]`).click() // open the menu
-      cy.get(`[data-testid="multi-select-clear"]`).click() // deselect previously selected tokens
-      cy.get(`[data-testid="menu-${columnId}"]`).should('not.exist') // clicking on clear closes the menu
-      cy.get(`[data-testid="multi-select-filter-${columnId}"]`).click() // open the menu again
-      cy.get(`[data-testid="menu-${columnId}"] [value="${symbol}"]`).click() // select the token
-      cy.get('body').click(0, 0) // close popover
-
-      getTableCellAssets().find(`[data-testid^="token-icon-${symbol}"]`).should('exist') // token might be hidden behind other tokens
-      cy.url().should('include', `assets_${type}_symbol=${encodeURIComponent(symbol)}`)
-
-      cy.get(`[data-testid="multi-select-filter-${columnId}"]`).click() // open the menu
-      cy.get(`[data-testid="multi-select-clear"]`).click() // deselect previously selected tokens
-      cy.url().should('not.include', `assets_${type}_symbol`)
-    }
-
-    /** Check if one or mor chains have been selected by checking url and table cell */
-    function checkChainSelection(...chains: Chain[]) {
-      cy.url().should('include', serializeListFilter(chains))
-      chains.forEach(chain => getTableCellAssets().find(`[data-testid="chain-icon-${chain}"]`).should('be.visible'))
-    }
-
-    /** Check the correct selection of the filter's grouped buttons by checking the url and a callback function */
-    function checkTableFilterButtonGroupSelection<T extends string>(
-      value: T,
-      type: string,
-      checkCallback: (value?: T) => void,
-    ) {
-      withFiltersPopover(() => cy.get(`[data-testid="table-filter-btn-market-${type}-${value}"]`).click())
-      cy.url().should('include', value)
-      checkCallback(value)
-    }
   })
 })
-
-function visitAndWait(
-  [width, height]: [number, number],
-  path = '/llamalend/ethereum/markets/',
-  options?: Partial<Cypress.VisitOptions>,
-) {
-  cy.viewport(width, height)
-  cy.visit(path, { ...LOAD_TIMEOUT, ...options })
-  cy.get('[data-testid="data-table"]', LOAD_TIMEOUT).should('be.visible')
-}
-
-function enableGraphColumn() {
-  cy.get(`[data-testid="line-graph-${MarketRateType.Borrow}"]`).should('not.exist')
-  cy.get(`[data-testid="btn-visibility-settings"]`).click()
-  cy.get(`[data-testid="visibility-toggle-borrowChart"]`).click()
-  cy.get(`[data-testid="line-graph-${MarketRateType.Borrow}"]`).should('exist')
-  cy.get('body').click(0, 0) // close popover
-}
-
-const typeFilterInput = (testId: string, value: number) =>
-  cy.get(`[data-testid="${testId}"]`).find('input[type="text"]').click().type('{selectAll}').type(`${value}`)
-
-/** Returns the midpoint between the min and max values for a column. */
-const getMiddleRange = <T extends Partial<Record<string, number>>>(data: T[], key: string) =>
-  mean([minBy(data, key)?.[key] ?? 0, maxBy(data, key)?.[key] ?? 0])
-
-/** Returns one random column paired with its middle range value. */
-const getOneColumnMiddleRangeValue = (
-  vaultData: Record<Chain, GetMarketsResponse>,
-  columns: readonly LlamaMarketColumnId[],
-) => {
-  const data = recordValues(vaultData).flatMap(({ data }) =>
-    data.map(
-      ({ borrowed_balance_usd, collateral_balance_usd, total_assets_usd, total_debt_usd, borrow_apr, max_ltv }) => ({
-        [LlamaMarketColumnId.LiquidityUsd]: total_assets_usd - total_debt_usd,
-        [LlamaMarketColumnId.Tvl]: borrowed_balance_usd + collateral_balance_usd + total_assets_usd - total_debt_usd,
-        [LlamaMarketColumnId.BorrowRate]: borrow_apr,
-        [LlamaMarketColumnId.UtilizationPercent]: total_assets_usd ? (100 * total_debt_usd) / total_assets_usd : 0,
-        [LlamaMarketColumnId.MaxLtv]: max_ltv,
-      }),
-    ),
-  )
-  const middleValues = fromEntries(columns.map(columnId => [columnId, getMiddleRange(data, columnId)] as const))
-  return oneOf(...recordEntries(middleValues))
-}
-
-const filterByMarketType = (size: [number, number], marketType: LlamaMarketType = LlamaMarketType.Lend) => {
-  const otherMarketType = oneOf(...recordValues(LlamaMarketType).filter(m => m !== marketType))
-  visitAndWait(size, `/llamalend/ethereum/markets?type=${marketType}`)
-  cy.url().should('include', `type=${marketType}`)
-  cy.get(`[data-testid="badge-market-type-${otherMarketType}"]`).should('not.exist')
-}
